requirements.txt
fastapi==0.115.5
pydantic==2.9.2
kafka-python==2.0.2
uvicorn==0.32.0
httpx==0.28.1